+++ include/download/api_parameter.h
@@ -1,7 +1,7 @@
 #ifndef API_PARAMETER_H
 #define API_PARAMETER_H
 
-#include "utils/json.hpp"
+#include "utils/nlohmann/json.hpp"
 #include <optional>
 #include <sstream>
 #include <string>
