+++ include/download/download_manager.h
@@ -1,7 +1,7 @@
 #ifndef DOWNLOAD_MANAGER_H
 #define DOWNLOAD_MANAGER_H
 
-#include "utils/json.hpp"
+#include "utils/nlohmann/json.hpp"
 #include <curl/curl.h>
 #include <fstream>
 #include <iostream>
